requirements.txt
git+git://github.com/django/django#egg=django # Dajngo 1.9
psycopg2==2.6

django-sampledatahelper==0.3

